Move AlignerClient and make it the surject base
